requirements.txt
ultralytics
tensorflow
Pillow
